requirements.txt
# PyPDF Toolbox GUI - Requirements
# Core PDF processing libraries
PyPDF2>=3.0.0
pypdf>=4.0.0
pymupdf>=1.23.0  # For PDF thumbnail generation (no external dependencies)
pdf2image>=1.16.0  # Alternative for thumbnails (requires poppler)
reportlab>=4.0.0

# OCR support
pytesseract>=0.3.10
pillow>=10.0.0

# GUI framework
# tkinter is included with Python standard library
tkinterdnd2>=0.3.0  # Drag and drop support

# Configuration
pyyaml>=6.0.0  # YAML configuration files

# Azure AI (optional - for AI-powered tools)
openai>=1.0.0  # Azure OpenAI client

# Additional utilities
tqdm>=4.65.0
